messaging_app/requirements.txt
djangorestframework
drf-nested-routers
djangorestframework-simplejwt